backend/api/pyproject.toml
[project]
name = "gitbacker-api"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.34",
    "sqlalchemy[asyncio]>=2.0",
    "asyncpg>=0.30",
    "alembic>=1.14",
    "pyjwt>=2.9",
    "bcrypt>=4.2",
    "pydantic[email]>=2.0",
    "celery>=5.4",
    "redis>=5.0",
    "python-dotenv>=1.0",
    "croniter>=6.0",
    "shared",
]

[project.optional-dependencies]
dev = ["pytest>=8.0", "pytest-asyncio>=0.24", "httpx>=0.27", "aiosqlite>=0.20"]

[tool.uv.sources]
shared = { path = "../shared", editable = true }

[tool.hatch.build.targets.wheel]
packages = ["app"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
pythonpath = ["."]
